pyproject.toml
[project]
name = "beacon"
version = "0.0.1"
description = "Multimodal geospatial event verification — claims to satellite imagery to grounded reports."
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "psycopg[binary]>=3.2",
    "claude-agent-sdk>=0.1",
    "huggingface-hub>=0.28",
    "pystac-client>=0.8",
    "planetary-computer>=1.0",
    "rasterio>=1.3",
    "pillow>=10",
    "shapely>=2.0",
    "structlog>=24",
    "typer>=0.12",
    "streamlit>=1.36",
    "python-dateutil>=2.9",
    "langgraph>=1.1.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "openpyxl>=3.1.5",
    "sentence-transformers>=5.4.1",
    "sentencepiece>=0.2.1",
    "langfuse>=2.50",
    "dspy-ai>=2.5",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/beacon"]

[project.scripts]
beacon = "beacon.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
